【master】增加自定义KVO实现属性值监听(使用runtime进行实现KVO) 【master】减少项目中不用的图片、不使用的项目代码文件 【master】减少警告(包括ARC警告问题，循环引用问题，代码合并问题) 【master】修改细节问题

+++ CollegePro/Class(首页)/Other/SZAddImage.m
@@ -177,7 +177,7 @@
 - (void)layoutSubviews
 {
     [super layoutSubviews];
-    int count = self.subviews.count;
+    int count = (int)self.subviews.count;
     CGFloat btnW = imageW;
     CGFloat btnH = imageH;
     int maxColumn = kMaxColumn > self.frame.size.width / imageW ? self.frame.size.width / imageW : kMaxColumn;
